+++ src/starlords/util/dialogControler/dialogValues/DialogValue_DialogData.java
@@ -1,9 +1,6 @@
 package starlords.util.dialogControler.dialogValues;
 
-import com.fs.starfarer.api.Global;
-import org.apache.log4j.Logger;
 import org.json.JSONObject;
-import starlords.lunaSettings.StoredSettings;
 import starlords.person.Lord;
 import starlords.plugins.LordInteractionDialogPluginImpl;
 
@@ -17,7 +14,7 @@ public class DialogValue_DialogData extends DialogValue_base{
     @Override
     public int value(Lord lord, Lord targetLord) {
         int rel = 0;
-        rel = LordInteractionDialogPluginImpl.DATA_HOLDER.getInteger(key);
+        rel = LordInteractionDialogPluginImpl.DATA_HOLDER.getDouble(key);
         return rel;
     }
